Data-Lake-d-Entreprise-main/requirements.txt
# Core Airflow
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.1.1
apache-airflow-providers-apache-hadoop==4.1.1

# Data Processing
pyspark==3.0.0

# Database
psycopg2-binary==2.9.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
